# Conflicts:
#	Guide/logging.markdown
#	ihp-hspec/default.nix
#	ihp-hspec/ihp-hspec.cabal
#	ihp-ide/IHP/IDE/ToolServer.hs
#	ihp-ide/exe/IHP/IDE/DevServer.hs
#	ihp-ide/ihp-ide.cabal
#	ihp-log/ihp-log.cabal
#	ihp-typed-sql/Test/Test/TypedSqlSpec.hs
#	ihp/IHP/Controller/Context.hs
#	ihp/IHP/FrameworkConfig.hs
#	ihp/IHP/Job/Runner/MainLoop.hs
#	ihp/IHP/Test/Mocking.hs
#	ihp/Test/Test/ControllerSupportSpec.hs
#	ihp/default.nix

+++ ihp/IHP/Pagination/ControllerFunctions.hs
@@ -102,8 +102,8 @@ paginateWithOptions options query = do
             }
 
     let results = query
-            |> limit pageSize
-            |> offset (offset' pageSize page)
+            |> limit (fromIntegral pageSize)
+            |> offset (fromIntegral (offset' pageSize page))
 
     pure
         ( results
